+++ src/util.ts
@@ -65,6 +65,7 @@ export const parseUrl = (
       .split("&")
       .map(pair => pair.split("="))
       .reduce((acc, val) => {
+        // acc[val[0]] = `${val[1]}`.valueOf();
         acc[val[0]] = val[1];
         return acc;
       }, {});