+++ changelog.md
@@ -33,6 +33,7 @@
 
 # 1.4.0
 
+- Changed the name of the 'functional classes' from Storage to Adapter, e.g. `StorageAmazonS3` became `AdapterAmazonS3`
 - Replaced `introspect()` by `getConfiguration():StorageConfig`, `getOptions():JSON` and `getType():string`
 - Added adapter class for BackBlaze B2
 - Made configuration is more strict and generic
@@ -41,4 +42,5 @@
 - Made slugify optional and turned it off by default on StorageLocal
 - Added API method `fileExists():Promise<boolean>`
 - In the options object or string non-existent keys or invalid values are no longer filtered out: it is up to the programmer to provide valid options
-- If no bucket name is provided the bucket name will always be an empty string "", not `undefined` or `null`.
+- If no bucket name is provided the bucket name will always be an empty string "", not `undefined` or `null`. Also when no bucket is selected `bucketName` will be "".
+- Adapter modules are only loaded when needed (using `require`)